requirements.txt
python-telegram-bot==13.7
gspread==4.0.1
oauth2client==4.1.3
